requirements.txt
flask==3.1.0
pandas>=2.0.0
reportlab>=4.0.0
werkzeug>=3.1.0
openpyxl>=3.1.0
pillow>=10.0.0
